requirements.txt
# Requires Python 3.10+
claude-agent-sdk>=0.1.53
python-dotenv>=1.0.0
anyio>=4.0.0
